Keep track of path information in GOSPEL to WhyML translation (wip)
Instead of only translating the identifier string of GOSPEL symbols,
now we propagate path information through the chain of translation
functions. Such an information is coded as a value of type `info`,
defined as follows:
```
  type info = {
    info_ls   : path Hls.t; (* lsymbol -> path    *)
    info_ts   : path Hts.t; (* tsymbol -> path    *)
    info_xs   : path Hxs.t; (* xsymbol -> path    *)
    info_path : path;       (* path = string list *)
  }
```
The `info_path` field stores the current path of module traversal.
Table info_{ls,ts,xs} associates each GOSPEL {ls,ts,xs}ymbol
to the corresponding path in the original `.mli` file. For instance,
in the following OCaml signature (example due to @pascutto and
@backtracking)
```
  module Make(X: sig type t end) : sig
    type t
    (*@ model f : X.t -> unit *)
  end
```
the ts_symbol `t` in functor argument `X` will be associated to path
`X :: Make :: []` (paths are stored in reverse order for performance
and convenience reasons). When translating `X.t` in the type of
`model f`, we are able to retrieve the correct path information,
which we did not in previous versions of the plugin. In this particular
case, the current path for translating `model f` is `Make :: []` and
the path for the `X.t` identifier is `X :: Make :: []`. It is thus
just a matter of reversing both paths and dropping the common prefix.
It must always be the case that the current path is a prefix of any
stored symbol's path.

For the moment, we only support translation with path information for
type symbols. We still need to implement the same approach for logical
and exception symbols.